Feature/users

+++ services_clothing_shop/src/main/java/mx/edu/utez/services_clothing_shop/service/user/UserService.java
@@ -2,6 +2,10 @@ package mx.edu.utez.services_clothing_shop.service.user;
 
 
 import mx.edu.utez.services_clothing_shop.controller.user.dto.*;
+import mx.edu.utez.services_clothing_shop.model.order.BeanOrder;
+import mx.edu.utez.services_clothing_shop.model.order_has_products.BeanOrderHasProducts;
+import mx.edu.utez.services_clothing_shop.model.order_has_products.IOrderHasProducts;
+import mx.edu.utez.services_clothing_shop.model.order_status.IOrderStatus;
 import mx.edu.utez.services_clothing_shop.model.role.BeanRole;
 import mx.edu.utez.services_clothing_shop.model.user_roles.IUserRoles;
 import mx.edu.utez.services_clothing_shop.service.email_service.EmailService;
@@ -26,13 +30,15 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
     private final EmailService emailService;
     private final IUserRoles userRolesRepository;
+    private final IOrderHasProducts orderHasProductsRepository;
 
-    public UserService(IUser userRepository, IRole roleRepository, PasswordEncoder passwordEncoder, EmailService emailService, IUserRoles userRolesRepository) {
+    public UserService(IUser userRepository, IRole roleRepository, PasswordEncoder passwordEncoder, EmailService emailService, IUserRoles userRolesRepository, IOrderHasProducts orderHasProductsRepository) {
         this.userRepository = userRepository;
         this.roleRepository = roleRepository;
         this.passwordEncoder = passwordEncoder;
         this.emailService = emailService;
         this.userRolesRepository = userRolesRepository;
+        this.orderHasProductsRepository = orderHasProductsRepository;
     }
 
 
@@ -105,6 +111,22 @@ public class UserService {
         if(user == null){
             throw new CustomException("user.email.not.found");
         }
+
+        List<BeanOrderHasProducts> orders = orderHasProductsRepository.findBySeller(payload.getEmail());
+        boolean hasOrderPending = false;
+
+        for (BeanOrderHasProducts order : orders) {
+            if(order.getStatus().getStatus().equals("Preparación") || order.getStatus().getStatus().equals("Enviado")){
+                hasOrderPending = true;
+                break;
+            }
+        }
+
+        if(hasOrderPending){
+            throw new CustomException("user.order.pending");
+        }
+
+
         user.setStatus(!user.isStatus());
 
         userRepository.saveAndFlush(user);
@@ -133,6 +155,22 @@ public class UserService {
         if(!passwordEncoder.matches(payload.getPassword(), user.getPassword())){
             throw new CustomException("user.password.incorrect");
         }
+
+        List<BeanOrderHasProducts> orders = orderHasProductsRepository.findBySeller(payload.getEmail());
+        boolean hasOrderPending = false;
+
+        for (BeanOrderHasProducts order : orders) {
+            if(order.getStatus().getStatus().equals("Preparación") || order.getStatus().getStatus().equals("Enviado")){
+                hasOrderPending = true;
+                break;
+            }
+        }
+
+        if(hasOrderPending){
+            throw new CustomException("user.own.order.pending");
+        }
+
+
 
         userRepository.deleteAccount(payload.getEmail());
 
@@ -281,6 +319,21 @@ public class UserService {
         }
 
 
+        List<BeanOrderHasProducts> orders = orderHasProductsRepository.findBySeller(payload.getEmail());
+        boolean hasOrderPending = false;
+
+        for (BeanOrderHasProducts order : orders) {
+            if(order.getStatus().getStatus().equals("Preparación") || order.getStatus().getStatus().equals("Enviado")){
+                hasOrderPending = true;
+                break;
+            }
+        }
+
+        if(hasOrderPending){
+            throw new CustomException("user.order.pending");
+        }
+
+
         userRepository.deleteAccount(payload.getEmail());
 
         emailService.sendEmail(user.getEmail(),
@@ -288,6 +341,7 @@ public class UserService {
                 "Tu cuenta ha sido eliminada por un administrador",
                 payload.getReazon(),
                 "Atentamente, "+payload.getAdmin());
+
         return true;
     }
 }
